Cleanup - CLI command extraction to individual files
